move functions from install_primitives.py to utils.py
